Szoftver architektúra fejlesztése, Supabase integráció és típusjavítások
- **Adatmodell és Adatbázis**: Az `Event` entitás azonosítójának típusa `int`-ről `long`-ra változott, kiegészítve GSON annotációkkal a Supabase kompatibilitás érdekében. Az adatbázis verziója frissítésre került.
- **Hálózati kommunikáció**: Retrofit, OkHttp és Gson függőségek hozzáadva; az `INTERNET` jogosultság bevezetve a manifestben.
- **Repository mintázat**: Az `EventRepository` bevezetése az adatkezelés absztrakciójához, lecserélve a közvetlen DAO hívásokat a Fragmentekben.
- **Navigáció és UI**:
    - Az `event_id` és `expense_id` argumentumok típusa `long`-ra módosult a navigációs gráfban.
    - Javított `MainActivity` navigációs logika az agresszívebb backstack kezelésért és a konzisztens menüállapotért.
    - Sötét mód támogatása és modernizált Material3 stílus az elszámolás és eseménykezelő felületeken.
- **Hibajavítások**:
    - Típuskonverziós hibák javítása az adapterekben és fragmentekben.
    - Üres állapotok és hibaüzenetek pontosabb kezelése az elszámolásnál.
    - `compileSdk` verzió fixálása.

+++ app/src/main/java/com/example/kifizeti_android/data/dao/ExpenseDao.java
@@ -23,7 +23,7 @@ public interface ExpenseDao {
     void delete(Expense expense);
 
     @Query("SELECT * FROM expenses WHERE eventId = :eventId")
-    List<Expense> getExpensesForEvent(int eventId);
+    List<Expense> getExpensesForEvent(long eventId);
 
     @Query("SELECT * FROM expenses WHERE id = :expenseId LIMIT 1")
     Expense getExpenseById(int expenseId);
